+++ controlplane/src/core/bufservices/PlatformService.ts
@@ -7,7 +7,6 @@ import { OrganizationEventName, PlatformEventName } from '@wundergraph/cosmo-con
 import { PlatformService } from '@wundergraph/cosmo-connect/dist/platform/v1/platform_connect';
 import {
   CheckFederatedGraphResponse,
-  CheckOperationUsageStats,
   CheckSubgraphSchemaResponse,
   CompositionError,
   CreateAPIKeyResponse,
@@ -95,6 +94,7 @@ import {
   InspectorOperationResult,
   SchemaUsageTrafficInspector,
   collectOperationUsageStats,
+  InspectorSchemaChange,
 } from '../services/SchemaUsageTrafficInspector.js';
 import Slack from '../services/Slack.js';
 import {
@@ -602,6 +602,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           };
         }
 
+        let isInspectable = true;
         const hasBreakingChanges = schemaChanges.breakingChanges.length > 0;
 
         await schemaCheckRepo.createSchemaCheckChanges({
@@ -631,11 +632,17 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         const inspectedOperations: InspectorOperationResult[] = [];
         const compositionErrors: PlainMessage<CompositionError>[] = [];
 
-        // For operations checks we only consider breaking changes
-        const inspectorChanges = trafficInspector.schemaChangesToInspectorChanges(
-          schemaChanges.breakingChanges,
-          storedBreakingChanges,
-        );
+        let inspectorChanges: InspectorSchemaChange[] = [];
+        try {
+          // For operations checks we only consider breaking changes
+          // This method will throw if the schema changes cannot be converted to inspector changes
+          inspectorChanges = trafficInspector.schemaChangesToInspectorChanges(
+            schemaChanges.breakingChanges,
+            storedBreakingChanges,
+          );
+        } catch {
+          isInspectable = false;
+        }
 
         for (const composition of result.compositions) {
           const graphConfig = await fedGraphRepo.getConfig(composition.id);
@@ -656,13 +663,14 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
             }
           }
 
-          // We don't collect operation usage when we have composition errors
-          if (composition.errors.length === 0 && inspectorChanges.inspectable && inspectorChanges.changes.length > 0) {
+          // We don't collect operation usage when we have composition errors or
+          // when we don't have any inspectable changes. That means any breaking change is really breaking
+          if (composition.errors.length === 0 && isInspectable && inspectorChanges.length > 0) {
             if (graphConfig.trafficCheckDays <= 0) {
               continue;
             }
 
-            const result = await trafficInspector.inspect(inspectorChanges.changes, {
+            const result = await trafficInspector.inspect(inspectorChanges, {
               daysToConsider: graphConfig.trafficCheckDays,
               federatedGraphId: composition.id,
               organizationId: authContext.organizationId,
@@ -691,9 +699,6 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           hasBreakingChanges,
         });
 
-        const operationUsageStats: PlainMessage<CheckOperationUsageStats> =
-          collectOperationUsageStats(inspectedOperations);
-
         if (req.gitInfo && opts.githubApp) {
           const githubRepo = new GitHubRepository(opts.db, opts.githubApp);
           await githubRepo.createCommitCheck({
@@ -715,7 +720,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           },
           breakingChanges: schemaChanges.breakingChanges,
           nonBreakingChanges: schemaChanges.nonBreakingChanges,
-          operationUsageStats,
+          operationUsageStats: isInspectable ? collectOperationUsageStats(inspectedOperations) : undefined,
           compositionErrors,
         };
       });